chore: migrate chaincode and delete old folder
