+++ asset-transfer-basic/META-INF/statedb/couchdb/indexes/indexOwner.json
